+++ logprep/util/getter.py
@@ -90,23 +90,28 @@ class FileGetter(Getter):
 
 @define(kw_only=True)
 class HttpGetter(Getter):
-    """get files from an api or simple web server.
+    """Get files from an api or simple web server.
 
     Matching string examples:
 
     * Simple http target: :code:`http://your.target/file.yml`
     * Simple https target: :code:`https://your.target/file.json`
 
-    if you want to use basic auth, then you have to set the environment variables
+    If you want to use basic auth, then you have to set the environment variables:
 
         * :code:`LOGPREP_CONFIG_AUTH_USERNAME=<your_username>`
         * :code:`LOGPREP_CONFIG_AUTH_PASSWORD=<your_password>`
 
-    if you want to use oauth, then you have to set the environment variables
+    If you want to use oauth, then you have to set the environment variables:
 
         * :code:`LOGPREP_CONFIG_AUTH_TOKEN=<your_token>`
         * :code:`LOGPREP_CONFIG_AUTH_METHOD=oauth`
 
+    If you want to load multiple resources from different sources you can set multiple auth tokens
+    by appending a strictly monotonously rising index to the variable. For example by using
+    :code:`LOGPREP_CONFIG_AUTH_TOKEN_0`, :code:`LOGPREP_CONFIG_AUTH_TOKEN_1`,
+    :code:`LOGPREP_CONFIG_AUTH_TOKEN_2`, etc. Logprep always tries to use all token until one is
+    successful.
     """
 
     _sessions: dict = {}