+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 ### Features
 
 ### Improvements
+* add possibility to define multiple oauth tokens by just appending an index to the env like
+`LOGPREP_CONFIG_AUTH_TOKEN_0`, `LOGPREP_CONFIG_AUTH_TOKEN_1`. The index has to be strictly
+monotonously rising.
 
 ### Bugfix
 * fix loading of configuration inside the `AutoRuleCorpusTester` for `logprep test integration`